requirements.txt
# 量化策略软件依赖

# 数据源
tushare>=1.2.89

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 可视化
matplotlib>=3.7.0
plotly>=5.15.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 日志
loguru>=0.7.0

# 进度条
tqdm>=4.65.0

# 可选：Jupyter 支持
# jupyterlab>=4.0.0
# ipywidgets>=8.0.0
